[modules] Move Vitamin and VitaminController
